Split compact command
